made BackButton its own file and removed unneccessary includes
