--- client/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:flutter_lints/flutter.yaml
