+++ client/lib/pages/generate_trail_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_svg/svg.dart';
+import '../widgets/back_button.dart';
 
 class GenerateTrail extends StatelessWidget{
 
@@ -14,7 +13,7 @@ class GenerateTrail extends StatelessWidget{
           children: [
             Row(
               children: [
-                BackButton().build(context),
+                GoBackButton().build(context),
               ],
             ), 
             Center(
@@ -35,26 +34,6 @@ class GenerateTrail extends StatelessWidget{
         )  
       ),
     );
-  }
-}
-
-class BackButton extends StatelessWidget {
-
-  @override 
-  Widget build(BuildContext context) {
-    return TextButton.icon(
-      onPressed: () {
-        Navigator.pop(context); 
-      }, 
-
-      icon: SvgPicture.asset('assets/images/arrow-sm-left-svgrepo-com (1).svg',
-        width: 40,
-        height: 40,
-        colorFilter: ColorFilter.mode(Colors.black, BlendMode.srcIn)
-      ),
-
-      label: Text(''), 
-    ); 
   }
 }
